Implement basic service instantiation

+++ src/gatekeeper/gatekeeper/api/services.py
@@ -1,27 +1,34 @@
 from typing import List
 
-import connexion
-from connexion.exceptions import ProblemException
 from mongoengine.errors import DoesNotExist
 
+import connexion
+from connexion.exceptions import ProblemException
+from gatekeeper.app import broker
 from gatekeeper.exceptions import DescriptorNotFoundError, ServiceNotFoundError
 from gatekeeper.models.descriptors import (Descriptor, DescriptorSnapshot,
                                            DescriptorType)
 from gatekeeper.models.services import Service
+from gatekeeper.util.mongoengine_custom_json import to_custom_dict, to_custom_json
+import json
 
 
 def getServices():
     return Service.objects()
 
 
-def getServiceById(id):
+def _getServiceByIdOrFail(id: str) -> Service:
     try:
         return Service.objects(id=id).get()
     except DoesNotExist:
         raise ServiceNotFoundError()
 
 
-def getReferencedDescriptors(descriptor: Descriptor) -> List[Descriptor]:
+def getServiceById(id):
+    return _getServiceByIdOrFail(id)
+
+
+def _getReferencedDescriptors(descriptor: Descriptor) -> List[Descriptor]:
     """
     Given a service descriptor, returns a list of all descriptors referenced by the service
     descriptor or raises a connexion `ProblemException` with a user-friendly message.
@@ -36,8 +43,6 @@ def getReferencedDescriptors(descriptor: Descriptor) -> List[Descriptor]:
             detail='{} does not specify any network functions.'.format(
                 descriptor.content)
         )
-
-        # TODO This would be ok if there were references to other services instead
 
     for function in descriptor.content.network_functions:
         try:
@@ -54,8 +59,6 @@ def getReferencedDescriptors(descriptor: Descriptor) -> List[Descriptor]:
                 referencedDescriptors.append(referencedDescriptor)
                 referencedDescriptorIds.add(referencedDescriptor.id)
 
-            # TODO Handle recursive service descriptors
-
         except DoesNotExist:
             raise ProblemException(
                 status=400,
@@ -70,7 +73,7 @@ def getReferencedDescriptors(descriptor: Descriptor) -> List[Descriptor]:
     return referencedDescriptors
 
 
-def createDescriptorSnapshot(descriptor: Descriptor) -> DescriptorSnapshot:
+def _createDescriptorSnapshot(descriptor: Descriptor) -> DescriptorSnapshot:
     return DescriptorSnapshot(**{field: descriptor[field] for field in descriptor})
 
 
@@ -89,7 +92,7 @@ def addService(body):
             )
 
         # Get referenced descriptors
-        referencedDescriptors = getReferencedDescriptors(rootDescriptor)
+        referencedDescriptors = _getReferencedDescriptors(rootDescriptor)
         allDescriptors = [rootDescriptor] + referencedDescriptors
 
         # Create and save service document
@@ -98,7 +101,7 @@ def addService(body):
             vendor=rootDescriptor.content.vendor,
             name=rootDescriptor.content.name,
             version=rootDescriptor.content.version,
-            descriptorSnapshots=[createDescriptorSnapshot(d) for d in allDescriptors]
+            descriptorSnapshots=[_createDescriptorSnapshot(d) for d in allDescriptors]
         )
         service.save()
         return service, 201
@@ -107,10 +110,62 @@ def addService(body):
 
 
 def deleteServiceById(id):
-    try:
-        service = Service.objects(id=id).get()
-        # TODO Fail if the service has instances
-        service.delete()
-        return service
-    except DoesNotExist:
-        raise ServiceNotFoundError()
+    service = _getServiceByIdOrFail(id)
+
+    instanceCount = len(service.instances)
+    if instanceCount > 0:
+        raise ProblemException(
+            title="Bad Request",
+            detail=("The service has {:d} instances that need to be stopped before it "
+                    "can be deleted.").format(instanceCount)
+        )
+
+    service.delete()
+    return service
+
+
+# Service instances
+
+def getServiceInstances(serviceId):
+    service = _getServiceByIdOrFail(serviceId)
+    return service.instances
+
+
+def instantiateService(serviceId):
+    service = _getServiceByIdOrFail(serviceId)
+
+    # Generate payload for instantiation message
+    payload = {
+        "ingresses": [],
+        "egresses": [],
+        "user_data": {
+            "customer": {
+                "email": "pishahang@gmail.com",
+                "phone": None,
+                "keys": {"public": None, "private": None}
+            },
+            "developer": {"email":  None, "phone": None}
+        }
+    }
+    vnfdCounter = 0
+    for descriptor in service.descriptorSnapshots:
+        descriptorContent = json.loads(to_custom_json(descriptor.content))
+        descriptorContent['uuid'] = descriptor.id  # For backwards compatibility with SLM
+        if descriptor.type == DescriptorType.SERVICE.value:
+            payload['COSD'] = descriptorContent
+        if descriptor.type == DescriptorType.OPENSTACK.value:
+            payload['VNFD{:d}'.format(vnfdCounter)] = descriptorContent
+        elif descriptor.type == DescriptorType.KUBERNETES.value:
+            payload['CSD{:d}'.format(vnfdCounter)] = descriptorContent
+        vnfdCounter += 1
+
+    # Send instantiation message
+    validationReply = broker.call_sync_simple("service.instances.create", msg=payload)
+    if validationReply["status"] == "ERROR":
+        raise ProblemException(
+            status=500,
+            title="Instantiation Failed",
+            detail=validationReply["error"]
+        )
+
+    # TODO register for notification and create service instance in database, return instance
